+++ src/RFC3339.ml
@@ -67,12 +67,12 @@ let pp_date_time ?frac_s () formatter (dt : Time.Date_time'.t) =
       in
       if frac_s = 0 then
         Fmt.pf formatter "%04d-%02d-%02dT%02d:%02d:%02d%s" dt.year
-          (human_int_of_month dt.month)
+          dt.month
           dt.day dt.hour dt.minute second tz_off
       else
         let divisor = get_divisor frac_s in
         Fmt.pf formatter "%04d-%02d-%02dT%02d:%02d:%02d.%0*d%s" dt.year
-          (human_int_of_month dt.month)
+          dt.month
           dt.day dt.hour dt.minute second frac_s (ns / divisor) tz_off
 
 let of_date_time ?frac_s (dt : Time.Date_time'.t) : string option =