+++ src/date_time_components.ml
@@ -8,21 +8,6 @@ type weekday =
   | `Sat
   ]
 
-type month =
-  [ `Jan
-  | `Feb
-  | `Mar
-  | `Apr
-  | `May
-  | `Jun
-  | `Jul
-  | `Aug
-  | `Sep
-  | `Oct
-  | `Nov
-  | `Dec
-  ]
-
 let tm_int_of_weekday (wday : weekday) : int =
   match wday with
   | `Sun -> 0
@@ -44,54 +29,6 @@ let weekday_of_tm_int (x : int) : weekday option =
   | 6 -> Some `Sat
   | _ -> None
 
-let tm_int_of_month (month : month) : int =
-  match month with
-  | `Jan -> 0
-  | `Feb -> 1
-  | `Mar -> 2
-  | `Apr -> 3
-  | `May -> 4
-  | `Jun -> 5
-  | `Jul -> 6
-  | `Aug -> 7
-  | `Sep -> 8
-  | `Oct -> 9
-  | `Nov -> 10
-  | `Dec -> 11
-
-let month_of_tm_int (x : int) : month option =
-  match x with
-  | 0 -> Some `Jan
-  | 1 -> Some `Feb
-  | 2 -> Some `Mar
-  | 3 -> Some `Apr
-  | 4 -> Some `May
-  | 5 -> Some `Jun
-  | 6 -> Some `Jul
-  | 7 -> Some `Aug
-  | 8 -> Some `Sep
-  | 9 -> Some `Oct
-  | 10 -> Some `Nov
-  | 11 -> Some `Dec
-  | _ -> None
-
-let human_int_of_month (month : month) : int = tm_int_of_month month + 1
-
-let month_of_human_int (x : int) : month option = month_of_tm_int (x - 1)
-
-let compare_month (m1 : month) (m2 : month) : int =
-  compare (tm_int_of_month m1) (tm_int_of_month m2)
-
-module Month_set = struct
-  include CCSet.Make (struct
-      type t = month
-
-      let compare = compare_month
-    end)
-
-  let to_seq x = x |> to_list |> CCList.to_seq
-end
-
 let compare_weekday (d1 : weekday) (d2 : weekday) : int =
   compare (tm_int_of_weekday d1) (tm_int_of_weekday d2)
 
@@ -105,9 +42,9 @@ module Weekday_set = struct
   let to_seq x = x |> to_list |> CCList.to_seq
 end
 
-let weekday_of_month_day ~(year : int) ~(month : month) ~(day : int) :
+let weekday_of_month_day ~(year : int) ~(month : int) ~(day : int) :
   weekday option =
-  Ptime.(of_date (year, human_int_of_month month, day))
+  Ptime.(of_date (year, month, day))
   |> CCOpt.map Ptime.weekday
 
 let is_leap_year ~year =
@@ -119,20 +56,21 @@ let is_leap_year ~year =
 
 let day_count_of_year ~year = if is_leap_year ~year then 366 else 365
 
-let day_count_of_month ~year ~(month : month) =
+let day_count_of_month ~year ~month =
   match month with
-  | `Jan -> 31
-  | `Feb -> if is_leap_year ~year then 29 else 28
-  | `Mar -> 31
-  | `Apr -> 30
-  | `May -> 31
-  | `Jun -> 30
-  | `Jul -> 31
-  | `Aug -> 31
-  | `Sep -> 30
-  | `Oct -> 31
-  | `Nov -> 30
-  | `Dec -> 31
+  | 1 -> 31
+  | 2 -> if is_leap_year ~year then 29 else 28
+  | 3 -> 31
+  | 4 -> 30
+  | 5 -> 31
+  | 6 -> 30
+  | 7 -> 31
+  | 8 -> 31
+  | 9 -> 30
+  | 10 -> 31
+  | 11 -> 30
+  | 12 -> 31
+  | _ -> invalid_arg "Invalid month"
 
 type tz_info = Time_zone.t * Duration.t option
 
@@ -163,15 +101,3 @@ let make_tz_info ?tz ?tz_offset () : (tz_info, tz_info_error) result =
 
 let tz_offset_of_tz_info ((tz, tz_offset) : tz_info) =
   match tz_offset with Some x -> Some x | None -> Time_zone.to_fixed_offset tz
-
-let next_ymd ~year ~month ~day : (int * month * int) option =
-  let day_count = day_count_of_month ~year ~month in
-  if 1 <= day && day <= day_count then
-    if day < day_count then Some (year, month, succ day)
-    else
-      let month_int = tm_int_of_month month in
-      let next_month_int = succ month_int in
-      let year = year + (next_month_int / 12) in
-      let month = CCOpt.get_exn @@ month_of_tm_int @@ (next_month_int mod 12) in
-      Some (year, month, 1)
-  else None