+++ src/printers.ml
@@ -67,10 +67,10 @@ module Format_string_parsers = struct
             >>= (fun x ->
                 return
                   (map_string_to_size_and_casing x
-                     (Time.full_string_of_month date_time.month)))
+                     (CCOpt.get_exn @@ Time.full_string_of_month date_time.month)))
                 <|> (padding
                      >>= fun padding ->
-                     return (pad_int padding (human_int_of_month date_time.month))));
+                     return (pad_int padding date_time.month)));
         (attempt (string "day:")
          >> padding
          >>= fun padding -> return (pad_int padding date_time.day));